requirements.txt
fastapi
uvicorn
gunicorn
google-generativeai>=0.6.0
